refactor: 패키지 util -> external로 변경
