+++ src/main/java/roomescape/config/interceptor/RoleCheckInterceptor.java
@@ -9,8 +9,8 @@ import roomescape.config.annotation.Authority;
 import roomescape.domain.Role;
 import roomescape.dto.business.AccessTokenContent;
 import roomescape.exception.ForbiddenException;
-import roomescape.utility.CookieUtility;
-import roomescape.utility.JwtTokenProvider;
+import roomescape.external.auth.CookieUtility;
+import roomescape.external.auth.JwtTokenProvider;
 
 public class RoleCheckInterceptor implements HandlerInterceptor {
