+++ src/main/java/roomescape/config/payment/PaymentConfig.java
@@ -7,8 +7,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.web.client.RestClient;
-import roomescape.utility.PaymentClient;
-import roomescape.utility.TossPaymentClient;
+import roomescape.external.payment.PaymentClient;
+import roomescape.external.payment.TossPaymentClient;
 
 @Configuration
 @Profile("!test")
