+++ src/main/java/roomescape/external/auth/CookieUtility.java
@@ -1,4 +1,4 @@
-package roomescape.utility;
+package roomescape.external.auth;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
