+++ src/main/java/roomescape/external/payment/PaymentClient.java
@@ -1,4 +1,4 @@
-package roomescape.utility;
+package roomescape.external.payment;
 
 import roomescape.domain.PaymentResult;
