+++ src/main/java/roomescape/external/payment/TossPaymentClient.java
@@ -1,4 +1,4 @@
-package roomescape.utility;
+package roomescape.external.payment;
 
 
 import com.fasterxml.jackson.databind.ObjectMapper;
