+++ src/test/java/roomescape/api/ThemeApiTest.java
@@ -27,7 +27,7 @@ import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.utility.JwtTokenProvider;
+import roomescape.external.auth.JwtTokenProvider;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 class ThemeApiTest {
