+++ src/test/java/roomescape/service/AuthServiceTest.java
@@ -17,7 +17,7 @@ import roomescape.dto.request.LoginRequest;
 import roomescape.dto.response.AccessTokenResponse;
 import roomescape.exception.LoginFailException;
 import roomescape.repository.MemberRepository;
-import roomescape.utility.JwtTokenProvider;
+import roomescape.external.auth.JwtTokenProvider;
 
 @DataJpaTest
 public class AuthServiceTest {
