+++ src/test/java/roomescape/utility/CookieUtilityTest.java
@@ -12,6 +12,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import roomescape.exception.NotFoundException;
+import roomescape.external.auth.CookieUtility;
 
 class CookieUtilityTest {
